+++ src/pages/Profile/AdminTab/AdminTab.tsx
@@ -1,15 +1,7 @@
 import { Department } from '../../../models/department'
-import {
-  TableColumn,
-  TableEdit,
-} from '../../../components/Table/TableEdit/TableEdit'
-import Input from '../../../components/UI/Input/Input'
-import { useEffect, useState } from 'react'
-import { Dropdown } from '../../../components/UI/Dropdown/Dropdown'
+import { TableEdit } from '../../../components/Table/TableEdit/TableEdit'
+import { useCallback, useEffect, useState } from 'react'
 import { commonEntities } from '../../../services/api/commonModule/commonEntities/commonEntities'
-import { Avatar } from '../../../components/UI/Avatar/Avatar'
-import { MaskedInput } from '../../../components/MaskedInput/MaskedInput'
-import { DateInput } from '../../../components/UI/DateInput/DateInput'
 import './AdminTab.scss'
 import { UserRole } from '../../../models/userRole'
 import { UserPosition } from '../../../models/userPoistion'
@@ -24,17 +16,7 @@ import { validateDateOfBirth } from '../../../validation/validateDateOfBirth'
 import { validateEmail } from '../../../validation/validateEmail'
 import { validatePhone } from '../../../validation/validatePhone'
 import { validateRolePositionDepartmentSelected } from '../../../validation/validateRolePositionDepartmentSelected'
-
-const sex = [
-  {
-    id: 1,
-    name: 'Мужской',
-  },
-  {
-    id: 2,
-    name: 'Женский',
-  },
-]
+import { getAdminTableColumns } from './getAdminTableColumns'
 
 export default function AdminTab() {
   const [users, setUsers] = useState<UserProfile[]>([])
@@ -148,267 +130,71 @@ export default function AdminTab() {
     }
   }
 
-  const handleRowChange = (
-    rowId: number,
-    key: keyof UserProfile,
-    value: any,
-  ) => {
-    setUsers(prev => {
-      const newUsers = prev.map(user =>
-        user.id === rowId ? { ...user, [key]: value } : user,
-      )
-
-      const updatedUser = newUsers.find(user => user.id === rowId)
-      setEditedUsers(prev => ({
-        ...prev,
-        [rowId]: updatedUser!,
-      }))
-
-      let error = ''
-      switch (key) {
-        case 'lastName':
-          error = validateName(value, false).error ?? ''
-          break
-        case 'firstName':
-          error = validateName(value, false).error ?? ''
-          break
-        case 'patronymic':
-          error = validateName(value, false).error ?? ''
-          break
-        case 'birthDate':
-          error = validateDateOfBirth(value, false).error ?? ''
-          break
-        case 'email':
-          error = validateEmail(value, false).error ?? ''
-          break
-        case 'phoneNumber':
-          error = validatePhone(value, false).error ?? ''
-          break
-
-        default:
-          error = ''
-      }
-
-      const roleSelected = updatedUser?.role ? [updatedUser.role] : []
-      const positionSelected = updatedUser?.position
-        ? [updatedUser.position]
-        : []
-      const departmentSelected = updatedUser?.department
-        ? [updatedUser.department]
-        : []
-
-      const relatedValidation = validateRolePositionDepartmentSelected({
-        roleSelected,
-        positionSelected,
-        departmentSelected,
-      })
-
-      setValidationErrors(prev => ({
-        ...prev,
-        [rowId]: {
-          ...(prev[rowId] || {}),
-          [key]: error,
-          role: relatedValidation.isValid
-            ? ''
-            : (relatedValidation.error ?? ''),
-          position: relatedValidation.isValid
-            ? ''
-            : (relatedValidation.error ?? ''),
-          department: relatedValidation.isValid
-            ? ''
-            : (relatedValidation.error ?? ''),
-        },
-      }))
-      return newUsers
-    })
+  const validateField = (key: keyof UserProfile, value: any) => {
+    switch (key) {
+      case 'lastName':
+      case 'firstName':
+      case 'patronymic':
+        return validateName(value, false).error ?? ''
+      case 'birthDate':
+        return validateDateOfBirth(value, false).error ?? ''
+      case 'email':
+        return validateEmail(value, false).error ?? ''
+      case 'phoneNumber':
+        return validatePhone(value, false).error ?? ''
+      default:
+        return ''
+    }
   }
 
-  const columns: TableColumn<UserProfile>[] = [
-    {
-      key: 'image',
-      title: '',
-      disableFilter: true,
-      cellRender: row => <Avatar src={row.image.url} size="small" />,
-    },
-    {
-      key: 'lastName',
-      title: 'Фамилия',
-      cellRender: (row, onChange, isDisabled) => (
-        <Input
-          value={row.lastName}
-          onChange={e => onChange(e.target.value)}
-          disabled={isDisabled}
-          placeholder="Фамилия"
-          errorText={validationErrors[row.id]?.lastName}
-        />
-      ),
-    },
-    {
-      key: 'firstName',
-      title: 'Имя',
-      cellRender: (row, onChange, isDisabled) => (
-        <Input
-          value={row.firstName}
-          onChange={e => onChange(e.target.value)}
-          disabled={isDisabled}
-          placeholder="Имя"
-          errorText={validationErrors[row.id]?.firstName}
-        />
-      ),
-    },
-    {
-      key: 'patronymic',
-      title: 'Отчество',
-      cellRender: (row, onChange, isDisabled) => (
-        <Input
-          value={row.patronymic}
-          onChange={e => onChange(e.target.value)}
-          disabled={isDisabled}
-          placeholder="Отчество"
-          errorText={validationErrors[row.id]?.patronymic}
-        />
-      ),
-    },
-    {
-      key: 'role',
-      title: 'Роль',
-      cellRender: (row, onChange, isDisabled) => (
-        <Dropdown
-          selectedItems={roles
-            .filter(role => role.id === row.role?.id)
-            .map(role => ({ id: role.id, text: role.name }))}
-          onChangeDropdown={selectedItem => {
-            if (selectedItem && !Array.isArray(selectedItem)) {
-              onChange(selectedItem)
-            } else {
-              onChange(null)
-            }
-          }}
-          disabled={isDisabled}
-          placeholder="Роль"
-          items={roles.map(role => ({ id: role.id, text: role.name }))}
-          errorText={validationErrors[row.id]?.role}
-        />
-      ),
-    },
+  const handleRowChange = useCallback(
+    (rowId: number, key: keyof UserProfile, value: any) => {
+      setUsers(prevUsers => {
+        const newUsers = prevUsers.map(u =>
+          u.id === rowId ? { ...u, [key]: value } : u,
+        )
+        const updatedUser = newUsers.find(u => u.id === rowId)!
 
-    {
-      key: 'department',
-      title: 'Кафедра',
-      cellRender: (row, onChange, isDisabled) => (
-        <Dropdown
-          selectedItems={departments
-            .filter(department => department.id === row.department?.id)
-            .map(department => ({ id: department.id, text: department.name }))}
-          onChangeDropdown={selectedItem => {
-            if (selectedItem && !Array.isArray(selectedItem)) {
-              onChange(selectedItem)
-            } else {
-              onChange(null)
-            }
-          }}
-          disabled={isDisabled}
-          placeholder="Кафедра"
-          items={departments.map(department => ({
-            id: department.id,
-            text: department.name,
-          }))}
-          errorText={validationErrors[row.id]?.department}
-        />
-      ),
+        const error = validateField(key, value)
+
+        const relatedValidation = validateRolePositionDepartmentSelected({
+          roleSelected: updatedUser.role ? [updatedUser.role] : [],
+          positionSelected: updatedUser.position ? [updatedUser.position] : [],
+          departmentSelected: updatedUser.department
+            ? [updatedUser.department]
+            : [],
+        })
+
+        setEditedUsers(prev => ({ ...prev, [rowId]: updatedUser }))
+        setValidationErrors(prev => ({
+          ...prev,
+          [rowId]: {
+            ...(prev[rowId] || {}),
+            [key]: error,
+            role: relatedValidation.isValid
+              ? ''
+              : (relatedValidation.error ?? ''),
+            position: relatedValidation.isValid
+              ? ''
+              : (relatedValidation.error ?? ''),
+            department: relatedValidation.isValid
+              ? ''
+              : (relatedValidation.error ?? ''),
+          },
+        }))
+
+        return newUsers
+      })
     },
-    {
-      key: 'position',
-      title: 'Должность',
-      cellRender: (row, onChange, isDisabled) => (
-        <Dropdown
-          selectedItems={positions
-            .filter(position => position.id === row.position?.id)
-            .map(position => ({ id: position.id, text: position.name }))}
-          onChangeDropdown={selectedItem => {
-            if (selectedItem && !Array.isArray(selectedItem)) {
-              onChange(selectedItem)
-            } else {
-              onChange(null)
-            }
-          }}
-          disabled={isDisabled}
-          placeholder="Должность"
-          items={positions.map(position => ({
-            id: position.id,
-            text: position.name,
-          }))}
-          errorText={validationErrors[row.id]?.position}
-        />
-      ),
-    },
-    {
-      key: 'email',
-      title: 'Почта',
-      cellRender: (row, onChange, isDisabled) => (
-        <Input
-          value={row.email}
-          onChange={e => onChange(e.target.value)}
-          disabled={isDisabled}
-          placeholder="Почта"
-          errorText={validationErrors[row.id]?.email}
-        />
-      ),
-    },
-    {
-      key: 'phoneNumber',
-      title: 'Телефон',
-      cellRender: (row, onChange, isDisabled) => (
-        <MaskedInput
-          mask="+7 (999) 999-99-99"
-          value={row.phoneNumber ?? ''}
-          onChange={e => onChange(e.target.value)}
-          placeholder="+7 (___) ___-__-__"
-          disabled={isDisabled}
-          errorText={validationErrors[row.id]?.phoneNumber}
-        />
-      ),
-    },
-    {
-      key: 'sex',
-      title: 'Пол',
-      cellRender: (row, onChange, isDisabled) => (
-        <Dropdown
-          selectedItems={sex
-            .filter(s => s.name === row.sex)
-            .map(s => ({ id: s.id, text: s.name }))}
-          onChangeDropdown={selectedItem => {
-            if (selectedItem && !Array.isArray(selectedItem)) {
-              onChange(selectedItem.text)
-            } else {
-              onChange(null)
-            }
-          }}
-          disabled={isDisabled}
-          placeholder="Пол"
-          items={sex.map(s => ({
-            id: s.id,
-            text: s.name,
-          }))}
-        />
-      ),
-    },
-    {
-      key: 'birthDate',
-      title: 'Дата рождения',
-      disableFilter: true,
-      cellRender: (row, onChange, isDisabled) => (
-        <DateInput
-          placeholder="Дата рождения"
-          value={row.birthDate}
-          disabled={isDisabled}
-          onChange={e => onChange(e.target.value)}
-          onDateSelect={date => onChange(date)}
-          errorText={validationErrors[row.id]?.birthDate}
-        />
-      ),
-    },
-  ]
+    [setUsers, setEditedUsers, setValidationErrors],
+  )
+
+  const columns = getAdminTableColumns({
+    roles,
+    positions,
+    departments,
+    validationErrors,
+  })
 
   return (
     <>
